integrate re-fetch user from api and loading icon
